Removed some stuff+worked pm map object loading
I've realized that having the entire project is kinda dumb, in a way.
Seeing as all my code is in the core project.  Anyway, I've trashed the
idea of just one large map when I was having object loading/clearing and
collision issues. So I've gone back to working on the various entity
types and a way to streamline the whole map loading process.